+++ source/appPlugins/ToolMilling.py
@@ -3169,7 +3169,9 @@ class ToolMilling(Excellon, AppTool):
                 dwell = tools_dict[tool_uid_key]['data']["tools_mill_dwell"]
                 dwelltime = tools_dict[tool_uid_key]['data']["tools_mill_dwelltime"]
                 laser_min_power = tools_dict[tool_uid_key]['data']["tools_mill_min_power"]
-                laser_on_code = tools_dict[tool_uid_key]['data']["tools_mill_laser_on"]
+                laser_on_code = tools_dict[tool_uid_key]['data'].get(
+                    "tools_mill_laser_on", self.app.options.get("tools_mill_laser_on", "M3")
+                )
                 pp_geometry_name = tools_dict[tool_uid_key]['data']["tools_mill_ppname_g"]
 
                 spindle_dir = self.app.options['tools_mill_spindledir']