+++ source/camlib.py
@@ -3876,7 +3876,11 @@ class CNCjob(Geometry):
         except KeyError:
             self.laser_min_power = 0.0
 
-        self.laser_on_code = tool_dict['tools_mill_laser_on']
+        try:
+            self.laser_on_code = tool_dict['tools_mill_laser_on']
+        except KeyError:
+            # older projects / Tcl cncjob did not store this key
+            self.laser_on_code = self.app.options.get("tools_mill_laser_on", "M3")
 
         try:
             self.spindlespeed = float(tool_dict['tools_mill_spindlespeed'])